+++ server/src/main/java/net/oneandone/stool/server/Main.java
@@ -43,7 +43,7 @@ public class Main {
 
         config = ServerConfiguration.load(); // it's loaded twice, but I don't have injection here ...
         System.setProperty("loglevel", config.loglevel);
-        System.setProperty("server.port", Integer.toString(config.portFirst));
+        System.setProperty("server.port", Integer.toString(config.port));
         SpringApplication.run(Main.class, args);
     }